Update search API server.js
There was a bug when searching for recipes. It only returned the ingredient searched, not the other ingredients in the recipe.

Changing the flow: Now it will return just the name of the recipe as a link to the rest of the recipe, rather than returning everything about every recipe that includes any ingredient searched.